+++ CHANGELOG.md
@@ -6,3 +6,10 @@ v0.2
 - added next block
 - attempt to randomize entry points for pieces
 - reduced height to 19 blocks
+
+v0.3
+- added visual candy to score
+- added support for left and right triggers
+- added double down function
+- can now restart the game
+- added difficulity increase over time